On Linux, use cout instead of printf in grep to enjoy better performance.
On linux this gives us about 30% performance gain on
  $ time jrep -R struct linux-3.12.6

On MacOS using cout gives a huge performance penalty. The above command becomes
about four times slower.
Some quick profiling shows that when using printf we spend about 30% of the time
in printf, but when using cout, about 80% is spent in output operations, most of
it in lock related operations.